fix: recording import bug, scratch import bug
